pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ldraney-gmail-mcp"
version = "0.1.0"
description = "MCP server for Gmail — read, triage, draft, and send across multiple accounts"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "mcp>=1.0",
    "ldraney-gmail-sdk>=0.1.0",
]

[tool.uv.sources]
ldraney-gmail-sdk = { path = "../gmail-sdk", editable = true }

[project.scripts]
ldraney-gmail-mcp = "gmail_mcp.server:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
